Cleaned up repo for terminal action-only repo.
