+++ src/actions/mod.rs
@@ -1,0 +1,4 @@
+// This module exposes terminal functions for Ansi and Windows Console.
+pub mod ansi;
+#[cfg(windows)]
+pub mod wincon;
