--- src/interface/mod.rs
@@ -1,8 +1,0 @@
-//! TODO: placeholder for specific components to build a TUI.
-
-
-mod alert;
-
-pub use {
-    alert::AlertBox,
-};
